[MonologBundle] Refactored the configuration to allow adding handlers
A subsequent config file does not overwrite the whole stack anymore. An
handler can now be redefined using the same name and will keep its place
in the stack. A new handler will be added at the bottom of the stack due
to the way the config are merged.

Handlers of the stack now have a priority (defaulting to 0) which is used
to allow to reorder them. This allow to add an handler at the top of the
stack by giving it a higher priority. Handlers with the same priority stay
in the order where they are defined in the config files.
As the merging strategy of the Config component could change it is
recommended to use the priority when adding a new handler in a subsequent
file even to put it at the bottom.

The support of the new BufferHandler (type: buffer) and RotatingFileHandler
(type: rotatingfile) was also added.