Adding a Swift MessageFactory for performance
Prior to this, in order for the main logger to be instantiated, if you
had a swiftmailer handler, ultimately a single Swift_Message would be
created, even if you didn't need it. That sounds minor, but due to Swift's
autoloading, it required quite a few extra classes, which added a small
amount of extra memory (~1MB) and extra CPU due to the auto-loader (~10-20ms).