Default swiftmailer transport was always used
The transport of the default mailer was always used, even if the monolog handler
has a different mailer setting. When using a non-default mailer, the correct transport
was never called and the monolog emails never sent.

Instead of assuming the default mailer transport, I now set the transport of the used
mailer.