Prepared MonologBundle for Symfony 4
In Symfony 3 the `public` parameter for service definitions is defaulted to `true`, while in Symfony 4 it is defaulted to `false`.
In order to keep MonologBundle working with both major versions, all we need to do is to explicitly set that parameter on all service definitions.

This Pull Request adds `public="true"` to all service definitions that didn't specify it.

Signed-off-by: Romain Neutron <imprec@gmail.com>